feat(ci): make golangci-lint blocking and expand linter set
The CI lint job was non-blocking (continue-on-error: true) and ran the
default linter set only. With CI no longer ignoring lint output, baseline
fixes applied throughout the codebase to keep main green:

- Clear all errcheck violations on fire-and-forget Exec/Scan calls. Each
  recovery path that updates instance/volume status now logs failures via
  zerolog instead of silently dropping the error.
- Strip a stale `argIdx++` after LIMIT clause and remove the unused
  quotaMu / projectQuotaMu fields from nova.Service (verified dead via
  repo-wide grep).
- Drop a duplicate scheduler.NewReconciler call that was overwritten on
  the next line (ineffassign).
- Reword a comment in sqlite_adapter.go that staticcheck SA9009 flagged
  as a misformatted compiler directive (`// go:embed` → "embed.FS").
- Replace strings.Replace(..., -1) with strings.ReplaceAll in
  keystone.substituteURLTemplates (gocritic wrapperFunc).
- Refactor cmd/migrate-l3/main.go to a run() helper so deferred Close
  actually executes (gocritic exitAfterDefer).
- Fix gocritic appendAssign in middleware/metrics.go by building a fresh
  bounds slice instead of appending to store.latencyBuckets.
- Drop a redundant int32 conversion in pkg/hypervisor/libvirt.go.

The expanded .golangci.yml turns on bodyclose, unconvert, gocritic, and
revive in addition to the defaults. misspell stays off because the
codebase mixes British-English comments with the US-English OpenStack
API surface and neither locale flag handles that. Tests/contract
helpers are excluded from the noisier linters.

CI status job now requires lint to succeed alongside build, test,
contract-tests, and compat-check.

Why: Trust-cleanup task 1.7 from the readiness audit. Silent error
returns and ignored lint findings made the codebase look healthier
than it was; making lint blocking surfaces regressions at PR time.

+++ pkg/hypervisor/libvirt.go
@@ -394,7 +394,7 @@ func (m *VMManager) getVMStateReal(ctx context.Context, vmUUID string) (string, 
 	}
 
 	// Convert libvirt state to OpenStack state
-	stateStr := libvirtStateToString(int32(state))
+	stateStr := libvirtStateToString(state)
 	return stateStr, int(state), nil
 }
